+++ app/(auth)/signup/(SignupSteps)/step3/page.tsx
@@ -1,0 +1,7 @@
+import StepThree from '.'
+
+const page = () => {
+  return <StepThree />
+}
+
+export default page
