--- app/(auth)/signup/(SignupSteps)/step4/page.tsx
@@ -1,7 +1,0 @@
-import StepFour from '.'
-
-const page = () => {
-  return <StepFour />
-}
-
-export default page
