+++ app/page.tsx
@@ -1,5 +1,5 @@
 import LandingPage from './LandingPage'
-// import LandingPage from './HomePage/page'
+
 
 export default function Home() {
   return (
